chore: separate type struct to new file
