+++ storage.go
@@ -314,66 +314,3 @@ func buildUrlWithOption(urlStr string, options UrlOptions) string {
 func removeEmptyFolderName(filePath string) string {
 	return regexp.MustCompile(`\/\/`).ReplaceAllString(filePath, "/")
 }
-
-type SortBy struct {
-	Column string `json:"column"`
-	Order  string `json:"order"`
-}
-
-type FileUploadResponse struct {
-	Key     string `json:"Key"`
-	Message string `json:"message"`
-	Data    []byte
-	Code    string `json:"statusCode"`
-	Error   string `json:"error"`
-}
-
-type SignedUrlResponse struct {
-	SignedURL string `json:"signedURL"`
-}
-
-type FileSearchOptions struct {
-	Limit         int    `json:"limit"`
-	Offset        int    `json:"offset"`
-	SortByOptions SortBy `json:"sortBy"`
-}
-
-type FileObject struct {
-	Name           string      `json:"name"`
-	BucketId       string      `json:"bucket_id"`
-	Owner          string      `json:"owner"`
-	Id             string      `json:"id"`
-	UpdatedAt      string      `json:"updated_at"`
-	CreatedAt      string      `json:"created_at"`
-	LastAccessedAt string      `json:"last_accessed_at"`
-	Metadata       interface{} `json:"metadata"`
-	Buckets        Bucket
-}
-
-type ListFileRequestBody struct {
-	Limit         int    `json:"limit"`
-	Offset        int    `json:"offset"`
-	SortByOptions SortBy `json:"sortBy"`
-	Prefix        string `json:"prefix"`
-}
-
-type TransformOptions struct {
-	Width   int    `json:"width"`
-	Height  int    `json:"height"`
-	Resize  string `json:"resize"`
-	Format  string `json:"format"`
-	Quality int    `json:"quality"`
-}
-
-type UrlOptions struct {
-	Transform *TransformOptions `json:"transform"`
-	Download  bool              `json:"download"`
-}
-
-type SignedUploadUrlResponse struct {
-	Url string `json:"url"`
-}
-
-type UploadToSignedUrlResponse struct {
-	Key string `json:"key"`
-}
